(PC-13220)[PRO] LostPassword as functional component
