+++ pro/src/components/pages/LostPassword/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LostPassword'
